# apps/backend/docs/swagger.yaml
basePath: /
definitions:
...
  internal_file.FileResult:
    properties:
      created_at:
        type: string
      extension:
        type: string
      id:
        type: string
      mime_type:
        type: string
      name:
        type: string
      owner_id:
        type: string
      parent_folder_id:
        type: string
      size:
        type: integer
      starred:
        type: boolean
      updated_at:
        type: string
    type: object
  internal_file.PreviewKind:
    enum:
    - image
    - video
    - audio
    - pdf
    - text
    - other
    type: string
    x-enum-varnames:
    - PreviewKindImage
    - PreviewKindVideo
    - PreviewKindAudio
    - PreviewKindPDF
    - PreviewKindText
    - PreviewKindOther
  internal_file.PreviewResult:
    properties:
      content:
        type: string
      kind:
        $ref: '#/definitions/internal_file.PreviewKind'
      mime_type:
        type: string
      name:
        type: string
      signed_url:
        type: string
      size:
        type: integer
    type: object
  internal_file.UploadResult:
    properties:
      created_at:
        type: string
      extension:
        type: string
      id:
        type: string
      mime_type:
        type: string
      name:
        type: string
      owner_id:
        type: string
      parent_folder_id:
        type: string
      size:
        type: integer
      updated_at:
        type: string
    type: object
  internal_file.moveRequest:
    properties:
      parent_folder_id:
        type: string
    type: object
  internal_file.renameRequest:
    properties:
      name:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - name
    type: object
  internal_file.starredRequest:
    properties:
      starred:
        type: boolean
    type: object
...
host: localhost:4000
info:
  contact: {}
  description: CriByte — self-hosted Google Drive clone
  title: CriByte API
  version: "1.0"
paths:
...
  /api/files:
    get:
      description: List files in a folder or search by name
      parameters:
      - description: Parent folder ID
        in: query
        name: folder_id
        type: string
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_file.FileResult'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - SessionAuth: []
      summary: List files
      tags:
      - files
  /api/files/{id}:
    get:
      description: Get metadata for a file by its ID
      parameters:
      - description: File ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_file.FileResult'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - SessionAuth: []
      summary: Get file metadata
      tags:
      - files
    patch:
      consumes:
      - application/json
      description: Rename a file. Auto-resolves name conflicts.
      parameters:
      - description: File ID
        in: path
        name: id
        required: true
        type: string
      - description: New name
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_file.renameRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_file.FileResult'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - SessionAuth: []
      summary: Rename a file
      tags:
      - files
  /api/files/{id}/download:
    get:
      description: Download file content as attachment
      parameters:
      - description: File ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - SessionAuth: []
      summary: Download a file
      tags:
      - files
  /api/files/{id}/move:
    post:
      consumes:
      - application/json
      description: Move a file to a different folder
      parameters:
      - description: File ID
        in: path
        name: id
        required: true
        type: string
      - description: Target folder
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_file.moveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_file.FileResult'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - SessionAuth: []
      summary: Move a file
      tags:
      - files
  /api/files/{id}/permanent-delete:
    post:
      description: Delete a file permanently from trash. File must be in trash first.
      parameters:
      - description: File ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - SessionAuth: []
      summary: Permanently delete a file
      tags:
      - files
  /api/files/{id}/preview:
    get:
      description: Get preview info (kind, mime_type, content for text files, signed
        URL for others)
      parameters:
      - description: File ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_file.PreviewResult'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - SessionAuth: []
      summary: Preview a file
      tags:
      - files
  /api/files/{id}/restore:
    post:
      description: Restore a trashed file to its original location
      parameters:
      - description: File ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_file.FileResult'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - SessionAuth: []
      summary: Restore a file from trash
      tags:
      - files
  /api/files/{id}/serve:
    get:
      description: Serve file content directly (supports Range requests for video/audio)
      parameters:
      - description: File ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - SessionAuth: []
      summary: Serve file content inline
      tags:
      - files
  /api/files/{id}/star:
    post:
      consumes:
      - application/json
      description: Toggle the starred/favorite status of a file
      parameters:
      - description: File ID
        in: path
        name: id
        required: true
        type: string
      - description: Starred status
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_file.starredRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_file.FileResult'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - SessionAuth: []
      summary: Star or unstar a file
      tags:
      - files
  /api/files/{id}/trash:
    delete:
      description: Soft-delete a file (move to trash)
      parameters:
      - description: File ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_file.FileResult'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - SessionAuth: []
      summary: Trash a file
      tags:
      - files
  /api/files/starred:
    get:
      description: List all starred files for the current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_file.FileResult'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - SessionAuth: []
      summary: List starred files
      tags:
      - files
  /api/files/trash:
    get:
      description: List all files in trash for the current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_file.FileResult'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - SessionAuth: []
      summary: List trashed files
      tags:
      - files
  /api/files/upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload a file via multipart form. Optionally specify parent_folder_id
        and custom name.
      parameters:
      - description: File to upload
        in: formData
        name: file
        required: true
        type: file
      - description: Parent folder ID
        in: formData
        name: parent_folder_id
        type: string
      - description: Custom file name
        in: formData
        name: name
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_file.UploadResult'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "413":
          description: Request Entity Too Large
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - SessionAuth: []
      summary: Upload a file
      tags:
      - files
...
schemes:
- http
- https
securityDefinitions:
  SessionAuth:
    in: cookie
    name: session_id
    type: apiKey
swagger: "2.0"